splitting functions and test into more logical files
